feat: add support for CRON schedules
Adds support for CRON schedules in addition to time intervals.

feat: allow combining multiple intervals and CRON schedules

Multiple `--interval` and `--cron` options can be combined for more flexibility.

BREAKING CHANGE: The `--server` flag is deprecated. Use `--interval` or `--cron` options instead. When not used with `--interval` or `--cron`, it will use a default of 5 minutes.